+++ packages/dm-core-plugins/src/form/components/Form.tsx
@@ -8,9 +8,9 @@ import {
 import { Button } from '@equinor/eds-core-react'
 import { FormProvider, useForm } from 'react-hook-form'
 import styled from 'styled-components'
-import { RegistryProvider } from './context/RegistryContext'
-import { AttributeList } from './fields/ObjectField'
-import { TFormProps } from './types'
+import { RegistryProvider } from '../context/RegistryContext'
+import { AttributeList } from '../fields/ObjectField'
+import { TFormProps } from '../types'
 
 const Wrapper = styled.div`
   max-width: 650px;
