+++ packages/dm-core-plugins/src/form/fields/ObjectField.test.tsx
@@ -6,7 +6,7 @@ import {
   waitFor,
 } from '@testing-library/react'
 import React from 'react'
-import { Form } from '../Form'
+import { Form } from '../components/Form'
 import { mockBlueprintGet, wrapper } from '../test-utils'
 import { TFormProps } from '../types'
